+++ iris_heap_lang/lib/atomic_heap.v
@@ -8,7 +8,7 @@ From iris.prelude Require Import options.
 (** A general logically atomic interface for a heap. All parameters are
 implicit, since it is expected that there is only one [heapGS_gen] in scope that
 could possibly apply. For example:
-  
+
   Context `{!heapGS_gen hlc Σ, !atomic_heap}.
 
 Or, for libraries that require later credits:
@@ -25,6 +25,17 @@ projections (e.g., either use [Local Definition alloc] or avoid the name [alloc]
 altogether), and do not register an instance -- just make it a [Definition] that
 others can register later.
 *)
+
+(** This is the namespace that atomic heap implementations may use for their
+invariants (the [heap_inv] field of [atomic_heap] below).
+
+We hardcode it since one should have only one instance of [atomic_heap], so
+having it as a parameter of this module is unnecessary. An alternative would be
+adding it as a field of the class. This however would make it impossible for
+the client to open its own invariants around heap operations, as they won't be
+able to prove disjointness between their namespaces and the heap's. *)
+Definition atomic_heapN := nroot .@ "atomic_heap".
+
 Class atomic_heap := AtomicHeap {
   (* -- operations -- *)
   alloc : val;
@@ -35,6 +46,19 @@ Class atomic_heap := AtomicHeap {
   (** * Ghost state *)
   (** The assumptions about [Σ], and the singleton [gname]s (if needed) *)
   atomic_heapGS : gFunctors → Type;
+  (** [heap_inv] is an invariant that is assumed to hold by all operations. It
+      can't be allocated through the interface. Instead, implementations should
+      provide an initialization lemma that builds [atomic_heapGS] and [heap_inv]
+      so that clients can obtain a closed proof (once they have chosen an
+      implementation).
+
+      The invariant has to be allocated together with [atomic_heapGS] since it
+      is likely to contain singleton ghost state. The initialization lemma is
+      not part of the interface as it can only be applied by providing an
+      implementation-specific [Σ : gFunctors]. *)
+  heap_inv `{!heapGS_gen hlc Σ} {H : atomic_heapGS Σ} : iProp Σ;
+  #[global] heap_inv_persistent `{!heapGS_gen hlc Σ} {H : atomic_heapGS Σ} ::
+    Persistent (heap_inv (H:=H));
   (* -- predicates -- *)
   pointsto `{!heapGS_gen hlc Σ} {H : atomic_heapGS Σ} (l : loc) (dq: dfrac) (v : val) : iProp Σ;
   (* -- pointsto properties -- *)
@@ -46,22 +70,33 @@ Class atomic_heap := AtomicHeap {
     Persistent (pointsto (H:=H) l DfracDiscarded v);
   #[global] pointsto_as_fractional `{!heapGS_gen hlc Σ} {H : atomic_heapGS Σ} l q v ::
     AsFractional (pointsto (H:=H) l (DfracOwn q) v) (λ q, pointsto (H:=H) l (DfracOwn q) v) q;
-  pointsto_agree `{!heapGS_gen hlc Σ} {H : atomic_heapGS Σ} l dq1 dq2 v1 v2 :
-    pointsto (H:=H) l dq1 v1 -∗ pointsto (H:=H) l dq2 v2 -∗ ⌜v1 = v2⌝;
+  #[global] pointsto_combine_sep_gives `{!heapGS_gen hlc Σ} {H : atomic_heapGS Σ}
+    l dq1 dq2 v1 v2 ::
+    CombineSepGives (pointsto (H:=H) l dq1 v1) (pointsto (H:=H) l dq2 v2)
+      ⌜✓ (dq1 ⋅ dq2) ∧ v1 = v2⌝;
+  #[global] pointsto_combine_as `{!heapGS_gen hlc Σ} {H : atomic_heapGS Σ}
+    l dq1 dq2 v1 v2 ::
+    CombineSepAs (pointsto (H:=H) l dq1 v1) (pointsto (H:=H) l dq2 v2)
+      (pointsto (H:=H) l (dq1 ⋅ dq2) v1) | 60; (* higher cost than Fractional *)
   pointsto_persist `{!heapGS_gen hlc Σ} {H : atomic_heapGS Σ} l dq v :
     pointsto (H:=H) l dq v ==∗ pointsto (H:=H) l DfracDiscarded v;
   (* -- operation specs -- *)
   alloc_spec `{!heapGS_gen hlc Σ} {H : atomic_heapGS Σ} (v : val) :
+    heap_inv (H:=H) -∗
     {{{ True }}} alloc v {{{ l, RET #l; pointsto (H:=H) l (DfracOwn 1) v }}};
   free_spec `{!heapGS_gen hlc Σ} {H : atomic_heapGS Σ} (l : loc) (v : val) :
+    heap_inv (H:=H) -∗
     {{{ pointsto (H:=H) l (DfracOwn 1) v }}} free #l {{{ l, RET #l; True }}};
   load_spec `{!heapGS_gen hlc Σ} {H : atomic_heapGS Σ} (l : loc) :
-    ⊢ <<{ ∀∀ (v : val) q, pointsto (H:=H) l q v }>>
-        load #l @ ∅
-      <<{ pointsto (H:=H) l q v | RET v }>>;
+    heap_inv (H:=H) -∗
+    <<{ ∀∀ (v : val) q, pointsto (H:=H) l q v }>>
+      load #l @ ↑atomic_heapN
+    <<{ pointsto (H:=H) l q v | RET v }>>;
   store_spec `{!heapGS_gen hlc Σ} {H : atomic_heapGS Σ} (l : loc) (w : val) :
-    ⊢ <<{ ∀∀ v, pointsto (H:=H) l (DfracOwn 1) v }>> store #l w @ ∅
-      <<{ pointsto (H:=H) l (DfracOwn 1) w | RET #() }>>;
+    heap_inv (H:=H) -∗
+    <<{ ∀∀ v, pointsto (H:=H) l (DfracOwn 1) v }>>
+      store #l w @ ↑atomic_heapN
+    <<{ pointsto (H:=H) l (DfracOwn 1) w | RET #() }>>;
   (* This spec is slightly weaker than it could be: It is sufficient for [w1]
   *or* [v] to be unboxed.  However, by writing it this way the [val_is_unboxed]
   is outside the atomic triple, which makes it much easier to use -- and the
@@ -70,10 +105,12 @@ Class atomic_heap := AtomicHeap {
   [destruct (decide (a = b))] and it will simplify in both places. *)
   cmpxchg_spec `{!heapGS_gen hlc Σ} {H : atomic_heapGS Σ} (l : loc) (w1 w2 : val) :
     val_is_unboxed w1 →
-    ⊢ <<{ ∀∀ v, pointsto (H:=H) l (DfracOwn 1) v }>> cmpxchg #l w1 w2 @ ∅
-      <<{ if decide (v = w1)
-          then pointsto (H:=H) l (DfracOwn 1) w2 else pointsto (H:=H) l (DfracOwn 1) v
-        | RET (v, #if decide (v = w1) then true else false) }>>;
+    heap_inv (H:=H) -∗
+    <<{ ∀∀ v, pointsto (H:=H) l (DfracOwn 1) v }>>
+      cmpxchg #l w1 w2 @ ↑atomic_heapN
+    <<{ if decide (v = w1)
+        then pointsto (H:=H) l (DfracOwn 1) w2 else pointsto (H:=H) l (DfracOwn 1) v
+      | RET (v, #if decide (v = w1) then true else false) }>>;
 }.
 
 Global Arguments alloc : simpl never.
@@ -112,31 +149,48 @@ Section derived.
 
   Import notation.
 
+  Lemma pointsto_agree l dq1 dq2 v1 v2 :
+    l ↦{dq1} v1 -∗ l ↦{dq2} v2 -∗ ⌜v1 = v2⌝.
+  Proof.
+    iIntros "Hl1 Hl2". by iCombine "Hl1" "Hl2" gives %[_ ->].
+  Qed.
+
+  Lemma pointsto_combine l dq1 dq2 v1 v2 :
+    l ↦{dq1} v1 -∗ l ↦{dq2} v2 -∗ l ↦{dq1 ⋅ dq2} v1 ∗ ⌜v1 = v2⌝.
+  Proof.
+    iIntros "Hl1 Hl2".
+    iDestruct (pointsto_agree with "[$][$]") as "->".
+    by iCombine "Hl1 Hl2" as "$".
+  Qed.
+
   Lemma cas_spec (l : loc) (w1 w2 : val) :
     val_is_unboxed w1 →
-    ⊢ <<{ ∀∀ v, pointsto l (DfracOwn 1) v }>>
-      CAS #l w1 w2 @ ∅
+    heap_inv -∗
+    <<{ ∀∀ v, pointsto l (DfracOwn 1) v }>>
+      CAS #l w1 w2 @ ↑atomic_heapN
     <<{ if decide (v = w1)
         then pointsto l (DfracOwn 1) w2 else pointsto l (DfracOwn 1) v
       | RET #if decide (v = w1) then true else false }>>.
   Proof.
-    iIntros (? Φ) "AU". awp_apply cmpxchg_spec; first done.
+    iIntros (?) "#Hheap %Φ AU".
+    awp_apply cmpxchg_spec; [done..|].
     iApply (aacc_aupd_commit with "AU"); first done.
     iIntros (v) "H↦". iAaccIntro with "H↦"; first by eauto with iFrame.
     iIntros "$ !> HΦ !>". wp_pures. done.
   Qed.
 
   Lemma faa_spec (l : loc) (i2 : Z) :
-    ⊢ <<{ ∀∀ i1 : Z, pointsto l (DfracOwn 1) #i1 }>>
-      FAA #l #i2 @ ∅
+    heap_inv -∗
+    <<{ ∀∀ i1 : Z, pointsto l (DfracOwn 1) #i1 }>>
+      FAA #l #i2 @ ↑atomic_heapN
     <<{ pointsto l (DfracOwn 1) #(i1 + i2) | RET #i1 }>>.
   Proof.
-    iIntros (Φ) "AU". rewrite /faa_atomic. iLöb as "IH".
-    wp_pures. awp_apply load_spec.
+    iIntros "#Hheap %Φ AU". rewrite /faa_atomic. iLöb as "IH".
+    wp_pures. awp_apply load_spec; first done.
     iApply (aacc_aupd_abort with "AU"); first done.
     iIntros (i1) "H↦". iAaccIntro with "H↦"; first by eauto with iFrame.
     iIntros "$ !> AU !>". wp_pures.
-    awp_apply cas_spec; first done.
+    awp_apply cas_spec; [done..|].
     iApply (aacc_aupd with "AU"); first done.
     iIntros (m) "Hl".
     iAaccIntro with "Hl"; first by eauto with iFrame.
@@ -145,7 +199,6 @@ Section derived.
     - iModIntro. iLeft. iFrame. iIntros "AU". iModIntro. wp_pure.
       by iApply "IH".
   Qed.
-
 End derived.
 
 (** Proof that the primitive physical operations of heap_lang satisfy said interface. *)
@@ -164,43 +217,46 @@ Section proof.
   Context `{!heapGS_gen hlc Σ}.
 
   Lemma primitive_alloc_spec (v : val) :
-    {{{ True }}} primitive_alloc v {{{ l, RET #l; l ↦ v }}}.
+    True -∗ {{{ True }}} primitive_alloc v {{{ l, RET #l; l ↦ v }}}.
   Proof.
-    iIntros (Φ) "_ HΦ". wp_lam. wp_alloc l. iApply "HΦ". done.
+    iIntros "_ %Φ !> _ HΦ". wp_lam. wp_alloc l. iApply "HΦ". done.
   Qed.
 
   Lemma primitive_free_spec (l : loc) (v : val) :
-    {{{ l ↦ v }}} primitive_free #l {{{ l, RET #l; True }}}.
+    True -∗ {{{ l ↦ v }}} primitive_free #l {{{ l, RET #l; True }}}.
   Proof.
-    iIntros (Φ) "Hl HΦ". wp_lam. wp_free. iApply "HΦ". done.
+    iIntros "_ %Φ !> Hl HΦ". wp_lam. wp_free. iApply "HΦ". done.
   Qed.
 
   Lemma primitive_load_spec (l : loc) :
-    ⊢ <<{ ∀∀ (v : val) q, l ↦{q} v }>> primitive_load #l @ ∅
-      <<{ l ↦{q} v | RET v }>>.
+    True -∗
+    <<{ ∀∀ (v : val) q, l ↦{q} v }>> primitive_load #l @ ↑atomic_heapN
+    <<{ l ↦{q} v | RET v }>>.
   Proof.
-    iIntros (Φ) "AU". wp_lam.
+    iIntros "_ %Φ AU". wp_lam.
     iMod "AU" as (v q) "[H↦ [_ Hclose]]".
     wp_load. iMod ("Hclose" with "H↦") as "HΦ". done.
   Qed.
 
   Lemma primitive_store_spec (l : loc) (w : val) :
-    ⊢ <<{ ∀∀ v, l ↦ v }>> primitive_store #l w @ ∅
-      <<{ l ↦ w | RET #() }>>.
+    True -∗
+    <<{ ∀∀ v, l ↦ v }>> primitive_store #l w @ ↑atomic_heapN
+    <<{ l ↦ w | RET #() }>>.
   Proof.
-    iIntros (Φ) "AU". wp_lam. wp_let.
+    iIntros "_ %Φ AU". wp_lam. wp_let.
     iMod "AU" as (v) "[H↦ [_ Hclose]]".
     wp_store. iMod ("Hclose" with "H↦") as "HΦ". done.
   Qed.
 
   Lemma primitive_cmpxchg_spec (l : loc) (w1 w2 : val) :
     val_is_unboxed w1 →
-    ⊢ <<{ ∀∀ (v : val), l ↦ v }>>
-      primitive_cmpxchg #l w1 w2 @ ∅
+    True -∗
+    <<{ ∀∀ (v : val), l ↦ v }>>
+      primitive_cmpxchg #l w1 w2 @ ↑atomic_heapN
     <<{ if decide (v = w1) then l ↦ w2 else l ↦ v
       | RET (v, #if decide (v = w1) then true else false) }>>.
   Proof.
-    iIntros (? Φ) "AU". wp_lam. wp_pures.
+    iIntros (?) "_ %Φ AU". wp_lam. wp_pures.
     iMod "AU" as (v) "[H↦ [_ Hclose]]".
     destruct (decide (v = w1)) as [Heq|Hne];
       [wp_cmpxchg_suc|wp_cmpxchg_fail];
@@ -209,13 +265,12 @@ Section proof.
 End proof.
 
 (* NOT an instance because users should choose explicitly to use it
-     (using [Explicit Instance]). *)
+     (using [Existing Instance]). *)
 Definition primitive_atomic_heap : atomic_heap :=
-  {| atomic_heapGS _ := TCTrue;
-     alloc_spec _ _ _ _ := primitive_alloc_spec;
-     free_spec _ _ _ _ := primitive_free_spec;
-     load_spec _ _ _ _ := primitive_load_spec;
-     store_spec _ _ _ _ := primitive_store_spec;
-     cmpxchg_spec _ _ _ _ := primitive_cmpxchg_spec;
-     pointsto_persist _ _ _ _ := primitive_laws.pointsto_persist;
-     pointsto_agree _ _ _ _ := primitive_laws.pointsto_agree |}.
+  {| atomic_heapGS _ := TCTrue
+   ; alloc_spec _ _ _ _ := primitive_alloc_spec
+   ; free_spec _ _ _ _ := primitive_free_spec
+   ; load_spec _ _ _ _ := primitive_load_spec
+   ; store_spec _ _ _ _ := primitive_store_spec
+   ; cmpxchg_spec _ _ _ _ := primitive_cmpxchg_spec
+   ; pointsto_persist _ _ _ _ := primitive_laws.pointsto_persist |}.